playground/python/response_processing/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorclass==2.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
docopt==0.6.2
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip-upgrader==1.4.15
propcache==0.2.1
pyarrow==19.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
terminaltables==3.1.10
tokenizers==0.21.0
torch==2.6.0
torchtext==0.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3